+++ Block/Adminhtml/System/Config/AutoKeyExchangeAdmin.php
@@ -24,17 +24,25 @@ class AutoKeyExchangeAdmin extends \Magento\Framework\View\Element\Template
     private $autokeyexchange;
 
     /**
-     * @param \Magento\Framework\View\Element\Template\Context $context
-     * @param \Amazon\Pay\Model\Config\AutoKeyExchange        $autokeyexchange
-     * @param array                                            $data
+     * @var KeyUpgrade
+     */
+    private $keyUpgrade;
+
+    /**
+     * @param \Magento\Framework\View\Element\Template\Context  $context
+     * @param \Amazon\Pay\Model\Config\AutoKeyExchange          $autokeyexchange
+     * @param \Amazon\Pay\Model\Config\KeyUpgrade               $keyUpgrade
+     * @param array                                             $data
      */
     public function __construct(
         \Magento\Framework\View\Element\Template\Context $context,
         \Amazon\Pay\Model\Config\AutoKeyExchange $autokeyexchange,
+        \Amazon\Pay\Model\Config\KeyUpgrade $keyUpgrade,
         array $data = []
     ) {
         parent::__construct($context, $data);
         $this->autokeyexchange = $autokeyexchange;
+        $this->keyUpgrade = $keyUpgrade;
     }
 
     /**
@@ -58,6 +66,26 @@ class AutoKeyExchangeAdmin extends \Magento\Framework\View\Element\Template
      */
     public function getCurrency()
     {
-        return $this->autokeyexchange->getCurrency();
+        $currency = $this->autokeyexchange->getCurrency();
+        if ($currency) {
+            $currency = strtoupper($currency);
+        }
+        return $currency;
+    }
+
+    /**
+     * Whether to display Key Upgrade or AKE
+     */
+    public function canUpgrade()
+    {
+        return $this->keyUpgrade->getMwsKeyForScope() && !$this->keyUpgrade->getExistingPublicKeyId();
+    }
+
+    /**
+     * Return Key Upgrade info
+     */
+    public function getKeyUpgradeConfig()
+    {
+        return json_encode($this->keyUpgrade->getJsonAmazonKeyUpgradeConfig());
     }
 }