+++ Model/CustomerLinkManagement.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright 2016 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright © Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License").
  * You may not use this file except in compliance with the License.
@@ -58,6 +58,8 @@ class CustomerLinkManagement implements \Amazon\Pay\Api\CustomerLinkManagementIn
     private $random;
 
     /**
+     * CustomerLinkManagement constructor
+     *
      * @param CustomerLinkRepositoryInterface $customerLinkRepository
      * @param CustomerLinkFactory             $customerLinkFactory
      * @param CustomerInterface               $customerInterface
@@ -82,7 +84,7 @@ class CustomerLinkManagement implements \Amazon\Pay\Api\CustomerLinkManagementIn
     }
 
     /**
-     * {@inheritdoc}
+     * @inheritdoc
      */
     public function getByCustomerId($customerId)
     {
@@ -90,14 +92,15 @@ class CustomerLinkManagement implements \Amazon\Pay\Api\CustomerLinkManagementIn
     }
 
     /**
-     * {@inheritdoc}
+     * @inheritdoc
      */
     public function create(AmazonCustomerInterface $amazonCustomer)
     {
         $customerData = $this->customerDataFactory->create();
+        $sanitizedNames = $this->getSanitizedNameData($amazonCustomer);
 
-        $customerData->setFirstname($amazonCustomer->getFirstName());
-        $customerData->setLastname($amazonCustomer->getLastName());
+        $customerData->setFirstname($sanitizedNames['first_name']);
+        $customerData->setLastname($sanitizedNames['last_name']);
         $customerData->setEmail($amazonCustomer->getEmail());
         $password = $this->random->getRandomString(64);
 
@@ -107,7 +110,7 @@ class CustomerLinkManagement implements \Amazon\Pay\Api\CustomerLinkManagementIn
     }
 
     /**
-     * {@inheritdoc}
+     * @inheritdoc
      */
     public function updateLink($customerId, $amazonId)
     {
@@ -119,5 +122,21 @@ class CustomerLinkManagement implements \Amazon\Pay\Api\CustomerLinkManagementIn
             ->setCustomerId($customerId);
 
         $this->customerLinkRepository->save($customerLink);
+    }
+
+    /**
+     * Remove special characters allowed in Amazon account names, but not Magento account names
+     *
+     * @param AmazonCustomerInterface $customer
+     * @return array
+     */
+    private function getSanitizedNameData($customer)
+    {
+        $pattern = '/([^\p{L}\p{M}\,\-\_\.\'\s\d]){1,255}+/u';
+
+        return [
+            'first_name' => trim(preg_replace($pattern, '', htmlspecialchars_decode($customer->getFirstname()))),
+            'last_name'  => trim(preg_replace($pattern, '', htmlspecialchars_decode($customer->getLastname())))
+        ];
     }
 }