+++ Model/Resolver/CheckoutSessionConfig.php
@@ -18,6 +18,8 @@ class CheckoutSessionConfig implements ResolverInterface
     private $checkoutSessionManagement;
 
     /**
+     * CheckoutSessionConfig constructor
+     *
      * @param CheckoutSessionManagement $checkoutSessionManagement
      */
     public function __construct(
@@ -27,8 +29,10 @@ class CheckoutSessionConfig implements ResolverInterface
     }
 
     /**
+     * Get config from CheckoutSessionManagement
+     *
      * @param Field $field
-     * @param $context
+     * @param ContextInterface $context
      * @param ResolveInfo $info
      * @param array|null $value
      * @param array|null $args
@@ -37,8 +41,9 @@ class CheckoutSessionConfig implements ResolverInterface
     public function resolve(Field $field, $context, ResolveInfo $info, array $value = null, array $args = null)
     {
         $cartId = $args['cartId'] ?? null;
+        $omitPayloads = $args['omitPayloads'] ?? false;
 
-        $response = $this->checkoutSessionManagement->getConfig($cartId);
+        $response = $this->checkoutSessionManagement->getConfig($cartId, $omitPayloads);
         return array_shift($response);
     }
 }