+++ view/frontend/web/js/amazon-button.js
@@ -16,26 +16,28 @@ define([
     'ko',
     'jquery',
     'Amazon_Pay/js/action/checkout-session-config-load',
-    'Amazon_Pay/js/action/checkout-session-button-payload-load',
     'Amazon_Pay/js/model/storage',
     'mage/url',
     'Amazon_Pay/js/amazon-checkout',
     'Magento_Customer/js/customer-data',
     'Magento_Checkout/js/model/payment/additional-validators',
     'mage/storage',
-    'Magento_Checkout/js/model/error-processor'
+    'Magento_Checkout/js/model/error-processor',
+    'Magento_Ui/js/model/messageList',
+    'Amazon_Pay/js/amazon-add-to-cart',
 ], function (
         ko,
         $,
         checkoutSessionConfigLoad,
-        buttonPayloadLoad,
         amazonStorage,
         url,
         amazonCheckout,
         customerData,
         additionalValidators,
         storage,
-        errorProcessor
+        errorProcessor,
+        globalMessageList,
+        amazonAddToCart
     ) {
     'use strict';
 
@@ -49,20 +51,45 @@ define([
         },
 
         drawing: false,
+        amazonPayButton: null,
+        currencyCode: null,
 
-        _loadButtonConfig: function (callback) {
+        _loadButtonConfig: function (callback, forceReload = false) {
             checkoutSessionConfigLoad(function (checkoutSessionConfig) {
                 if (!$.isEmptyObject(checkoutSessionConfig)) {
-                    callback({
+                    var payload = checkoutSessionConfig['checkout_payload'];
+                    var signature = checkoutSessionConfig['checkout_signature'];
+
+                    if (this.buttonType === 'PayNow') {
+                        payload = checkoutSessionConfig['paynow_payload'];
+                        signature = checkoutSessionConfig['paynow_signature'];
+                    }
+
+                    self.currencyCode = checkoutSessionConfig['currency'];
+
+                    var buttonConfig = {
                         merchantId: checkoutSessionConfig['merchant_id'],
                         publicKeyId: checkoutSessionConfig['public_key_id'],
-                        ledgerCurrency: checkoutSessionConfig['currency'],
+                        ledgerCurrency: self.currencyCode,
                         sandbox: checkoutSessionConfig['sandbox'],
                         checkoutLanguage: checkoutSessionConfig['language'],
-                        productType: this._isPayOnly() ? 'PayOnly' : 'PayAndShip',
+                        productType: this._isPayOnly(checkoutSessionConfig['pay_only']) ? 'PayOnly' : 'PayAndShip',
                         placement: this.options.placement,
-                        buttonColor: checkoutSessionConfig['button_color']
-                    });
+                        buttonColor: checkoutSessionConfig['button_color'],
+                        createCheckoutSessionConfig: {
+                            payloadJSON: payload,
+                            signature: signature,
+                            publicKeyId: checkoutSessionConfig['public_key_id'],
+                        }
+                    };
+
+                    if (this._shouldUseEstimatedAmount()
+                        && !(JSON.parse(checkoutSessionConfig.checkout_payload).recurringMetadata)
+                    ) {
+                        buttonConfig.estimatedOrderAmount = this._getEstimatedAmount();
+                    }
+
+                    callback(buttonConfig);
 
                     if (this.options.placement !== "Checkout") {
                         $(this.options.hideIfUnavailable).show();
@@ -70,41 +97,33 @@ define([
                 } else {
                     $(this.options.hideIfUnavailable).hide();
                 }
-            }.bind(this));
+            }.bind(this), forceReload);
         },
 
-        _loadInitCheckoutPayload: function (callback, payloadType) {
-            checkoutSessionConfigLoad(function (checkoutSessionConfig) {
-                buttonPayloadLoad(function (buttonPayload) {
-                    callback({
-                        createCheckoutSessionConfig: {
-                            payloadJSON: buttonPayload[0],
-                            signature: buttonPayload[1],
-                            publicKeyId: checkoutSessionConfig['public_key_id']
-                        }
-                    });
-                }, payloadType);
-            });
+        _getEstimatedAmount: function () {
+            var subtotal = (parseFloat(customerData.get('cart')().subtotalAmount) || 0).toFixed(2);
+
+            if (self.currencyCode === 'JPY') {
+                subtotal = parseFloat(subtotal).toFixed(0);
+            }
+
+            return {
+                amount: subtotal,
+                currencyCode: self.currencyCode
+            };
         },
 
         /**
+         * @param {boolean} isCheckoutSessionPayOnly
          * @returns {boolean}
          * @private
          */
-        _isPayOnly: function () {
-            var cartData = customerData.get('cart');
-
-            // No cart data yet or cart is empty, for the pdp button
-            if (typeof cartData().amzn_pay_only === 'undefined' || parseInt(cartData().summary_count) === 0) {
-                return this.options.payOnly
+         _isPayOnly: function (isCheckoutSessionPayOnly) {
+            var result = isCheckoutSessionPayOnly;
+            if (result && this.options.payOnly !== null) {
+                result = this.options.payOnly;
             }
-
-            // Check if cart has items and it's the pdp button
-            if (parseInt(cartData().summary_count) > 0 && this.options.payOnly !== null) {
-                return cartData().amzn_pay_only && this.options.payOnly;
-            }
-
-            return cartData().amzn_pay_only;
+            return result;
         },
 
         /**
@@ -124,7 +143,7 @@ define([
 
         /**
          * Draw button
-         **/
+         */
         _draw: function () {
             var self = this;
 
@@ -137,59 +156,104 @@ define([
                     $buttonContainer.empty().append($buttonRoot);
 
                     this._loadButtonConfig(function (buttonConfig) {
+                        // remove session config to decouple button, allowing onclick adjustment
+                        if(self._isAmazonPayShownAsPaymentMethod() || self.options.placement === 'Product') {
+                            delete buttonConfig.createCheckoutSessionConfig;
+                        }
+
                         try {
-                            var amazonPayButton = amazon.Pay.renderButton('#' + $buttonRoot.empty().removeUniqueId().uniqueId().attr('id'), buttonConfig);
+                            self.amazonPayButton = amazon.Pay.renderButton('#' + $buttonRoot.empty().removeUniqueId().uniqueId().attr('id'), buttonConfig);
                         } catch (e) {
                             console.log('Amazon Pay button render error: ' + e);
                             return;
                         }
-                        amazonPayButton.onClick(function() {
-                            if (self.buttonType === 'PayNow' && !additionalValidators.validate()) {
-                                return false;
-                            }
-                            //This is for compatibility with Iosc. We need to update the customer's Magento session before getting the final config and payload
-                            if (self.buttonType === 'PayNow' && self.options.isIosc()) {
-                                storage.post(
-                                    'checkout/onepage/update',
-                                    "{}",
-                                    false
-                                ).done(
-                                    function (response) {
-                                        if (!response.error) {
-                                            self._initCheckout(amazonPayButton);
-                                        } else {
-                                            errorProcessor.process(response);
-                                        }
-                                    }
-                                ).fail(
-                                    function (response) {
-                                        errorProcessor.process(response);
-                                    }
-                                );
-                            }else{
-                                self._initCheckout(amazonPayButton);
-                            }
-                        });
+
+                        if (self._isAmazonPayShownAsPaymentMethod()) {
+                            self.amazonPayButton.onClick(function() {
+                                if (!additionalValidators.validate()) {
+                                    return false;
+                                }
+                                self._initCheckout();
+                            });
+                        } else if (self.options.placement === 'Product') {
+                            self.amazonPayButton.onClick(function() {
+                                amazonAddToCart.execute().then(function() {
+                                    self._initCheckout();
+                                });
+                            });
+                        }
 
                         $('.amazon-button-container .field-tooltip').fadeIn();
                         self.drawing = false;
+
+                        if (self.buttonType === 'PayNow' && self._isPayOnly()) {
+                            customerData.get('checkout-data').subscribe(function (checkoutData) {
+                                const opacity = checkoutData.selectedBillingAddress ? 1 : 0.5;
+
+                                const shadow = $('.amazon-checkout-button > div')[0].shadowRoot;
+                                $(shadow).find('.amazonpay-button-view1').css('opacity', opacity);
+                            });
+                        }
                     });
                 }, this);
             }
         },
 
-        _initCheckout: function (amazonPayButton) {
-            var payloadType = this.buttonType ?
-                'paynow' :
-                'checkout';
-            this._loadInitCheckoutPayload(function (initCheckoutPayload) {
-                amazonPayButton.initCheckout(initCheckoutPayload);
-            }, payloadType);
+        // PayNow indicating this is an APB button
+        _isAmazonPayShownAsPaymentMethod: function () {
+            return (this.buttonType === 'PayNow');
+        },
+
+        _initCheckout: function () {
+            if( this.options.isIosc()) {
+                this._initOneStepCheckout();
+            } else {
+                this._initAmazonCheckout();
+            }
+        },
+
+        _initOneStepCheckout: function () {
+            //This is for compatibility with Iosc. We need to update the customer's Magento session before getting the final config and payload
+            storage.post(
+                'checkout/onepage/update',
+                "{}",
+                false
+            ).done(
+                function (response) {
+                    if (!response.error) {
+                        self._initAmazonCheckout();
+                    } else {
+                        errorProcessor.process(response);
+                    }
+                }
+            ).fail(
+                function (response) {
+                    errorProcessor.process(response);
+                }
+            );
+        },
+
+        _initAmazonCheckout: function () {
+            var self = this;
+
+            if (self.buttonType === 'PayNow' && self._isPayOnly()) {
+                if (!customerData.get('checkout-data')().selectedBillingAddress) {
+                    return;
+                } else {
+                    var setBillingAddressAction = require('Magento_Checkout/js/action/set-billing-address');
+                    setBillingAddressAction(globalMessageList);
+                }
+            }
+
+            this._loadButtonConfig(function (buttonConfig) {
+                var initConfig = {createCheckoutSessionConfig: buttonConfig.createCheckoutSessionConfig};
+                self.amazonPayButton.initCheckout(initConfig);
+            }, true);
             customerData.invalidate('*');
         },
 
         /**
-         * Redraw button if needed
+         * Update button if needed
          **/
         _subscribeToCartUpdates: function () {
             var self = this;
@@ -197,18 +261,47 @@ define([
             amazonCheckout.withAmazonCheckout(function (amazon, args) {
                 var cartData = customerData.get('cart');
                 cartData.subscribe(function (updatedCart) {
-                    if (!$(self.options.hideIfUnavailable).first().is(':visible')) {
-                        self._draw();
+                    if (self.amazonPayButton && self.buttonType !== 'PayNow') {
+                        var hasSubscription = false;
+                        var isSubscriptionOption = (option) => option.label && option.label === amazonStorage.getSubscriptionLabel();
+
+                        // If any cart item has a subscription option, we should redraw the button to exclude estimatedOrderAmount
+                        updatedCart.items.forEach((item) => {
+                            if (item.options.length && item.options.some(isSubscriptionOption)) {
+                                hasSubscription = true;
+                                return;
+                            }
+                        });
+
+                        if (hasSubscription) {
+                            self._draw();
+                        } else {
+                            if (self.options.placement === 'Cart') {
+                                delete self.amazonPayButton;
+                            }
+
+                            if (self.amazonPayButton
+                                && self._shouldUseEstimatedAmount()
+                                && updatedCart.summary_count !== 0
+                            ) {
+                                self.amazonPayButton.updateButtonInfo(self._getEstimatedAmount());
+                            }
+                        }
                     }
                 });
             });
+        },
+
+        _shouldUseEstimatedAmount: function () {
+            return this.options.buttonType !== 'PayNow'
+                && this.options.placement !== 'Product'
+                && !amazonStorage.isMulticurrencyEnabled;
         },
 
         click: function () {
             this.element.children().first().trigger('click');
         }
     });
-
 
     var cart = customerData.get('cart'),
         customer = customerData.get('customer'),