+++ view/frontend/web/js/model/storage.js
@@ -31,11 +31,13 @@ define([
 
     var isLwaEnabled = amazonPayConfig.getValue('is_lwa_enabled');
     var isGuestCheckoutEnabled = amazonPayConfig.getValue('is_guest_checkout_enabled');
+    var isMulticurrencyEnabled = amazonPayConfig.getValue('is_multicurrency_enabled');
 
     return {
         isEnabled: isEnabled,
         isLwaEnabled: isLwaEnabled,
         isGuestCheckoutEnabled: isGuestCheckoutEnabled,
+        isMulticurrencyEnabled: isMulticurrencyEnabled,
 
         /**
          * Is checkout using Amazon Pay?
@@ -78,6 +80,13 @@ define([
          */
         getRegion: function() {
             return amazonPayConfig.getValue('region');
+        },
+
+        /**
+         * Return label for subscription option defined in ParadoxLabs config
+         */
+        getSubscriptionLabel: function () {
+            return amazonPayConfig.getValue('subscription_label');
         }
     };
 });