+++ CHANGELOG.md
@@ -1,5 +1,47 @@
 # Change Log
 
+## 5.16.1
+* Remove platform/module version information from button payloads
+* Fixed admin display bug regarding upgrading legacy keys when a CV2 public key ID already exists
+* Fixed invoice comment noise with Automatic Authorization Mode
+
+## 5.16.0
+* Added support for subscriptions, requires the ParadoxLabs Adaptive Subscription module
+* Added platformid for APB payload
+* Fixed phpcs warnings for php 8
+* Fixed issue with installing the module when there is no existing CV1 config
+* Fixed issue when an Allowed Country was somehow saved with an empty string
+
+## 5.15.0
+* Added Key Upgrade feature (automatically upgrade CV1 keys to CV2 if updating from legacy module)
+* Fixed backward compatibility with Zend availability
+* Fixed render issue with multiple Amazon Sign In buttons on the same page
+
+## 5.14.3
+* Fixed PHP 7 compatibility
+
+## 5.14.2
+* Removed usage of Zend libraries and updated requirements for PHP 8.2/Magento 2.4.6 compatibility
+* Fixed issue where coupon usage is not properly decremented when placing order fails
+
+## 5.14.1
+* Changed how buttons are rendered for compatibility with estimated order amount feature
+* Removed estimated order amount from PDP button
+
+## 5.14.0
+* Added configurable options for checkout and signin cancel return urls
+* Added estimated order amount to the button payload
+* Fixed issue with ExceptionLogger using a removed method
+* Fixed issue with customer names having characters that Magento doesn’t allow
+* Fixed issue with APB and virtual orders when taxes are applied
+* Improved compatibility with b2b modules and address display
+* Updated config to not show Auto Key Exchange button for JPY as it is not supported
+
+## 5.13.1
+* Fixed issue with invalid array reference if a transaction doesn’t have a charge permission
+* Fixed issue with GraphQL config query supporting omitPayloads
+* Fixed issue with switching stores to refresh the Amazon Pay button config
+
 ## 5.13.0
 * Added Graphql support
 * Added endpoints to fetch individual config types